Added initial difficulty setting
I have now implemented the ability to set the difficulty before playing
the game.

Elementary Difficulty: adding two random numbers from 0 to 10
Middle School Difficulty: adding two random numbers from 0 to 20
High School Difficulty: adding two random numbers from -20 to 20

NOTE: This is not an official release update to the game, it is just an
update to the code. In time, this code will be built and put in an
executable jar file and uploaded to SourceForge as an official release.