Added command /help
This commit adds the /help command, which displays all the commands
available, their arguments, and what they do. Also added was the use of
the VARIABLE_STATE Field. This is so that the user cannot use any
commands until he/she enters in a Difficulty Level, and if I add more
functionality, there will be more states.