Added logging capabilities
When the program runs on a Mac, the directory
Users/[username]/Library/Application Support/mathquizgame/ is created
if it doesn't already exist, and log.txt is created in that directory
if the file doesn't exist. Whatever the program outputs to the user or
the user types in, it will be appended to the log.txt file. I need to
make a version for Windows, so when the program runs on Windows, it
will create the file in Program Files or something.