Converted Commands to actual meathods
decreased size of somethingHappened() dramatically by moving functions
in if statements to meathods in the main class. This way, commands can
be accessed by other places in the program.

Also, the import block was cleaned out of all the unused stuff, and
command list was updated in the /help command.