Custom Difficulty Level Added
The source code has been redesigned into three different classes as
opposed to having two classes in the latest version. The class with the
main method, MathQuizGame, handles all GUI and user input, and when the
user enters in an answer, MathQuizGame sends that answer to
MathOperator, which addresses all the math stuff.

Here's what changed in MathQuizGame:
*Difficulty Fields removed
*Fixed wrong way of expressing the directory for the game history
*Enabled findCommandArgument() to find more than one argument
*newNumbers() is not needed. Will be removed from code in next commit
*nextQuestion() is not needed. Will be removed from code in next commit
*Changed way of starting a new game
*Enabled commands to be used during Difficulty Changing State