Added command /history and fixed indexArrayThing not resetting
The new command /history displays everything you entered in, or at
least it tries to. It's not working and I think it's because of the for
loop statement. I'll work on that later.

Also, when you type something besides the up or down arrow,
indexArrayThing gets reset. You still have to press the up arrow twice
after it's reset, though, and that needs to be fixed.