Deleted unnecessary duplicate else if statements and changed * to x for Multiply
I deleted most of the if statements under the statements that determine
the state so that most of the commands can be used in any state.
Earlier, if I needed the commands to be available in any state, I had
to copy them. Now there’s only one.

Also, the character used to illustrate multiplication has turned
globaly to x instead of *

Renamed character finals in MathOperator.