Added padding around the debug screen, more debug improvements, NullPointerException upon launch bug fixed.
Not a good idea to use the object you’re constructing is it? The game
wouldn’t launch correctly because EnterText() was accessing
MathQuizGame.frame, which was at the time, when EnterText() was called,
being initialized, and the program threw a NullPointerException.

This bug was related to my conversion to non-static fields in
MathQuizGame.