Fixed argument parsing for custom difficulty level.
parseArgs() would return something crazy if the command does’t start
with a “/”. It still does, but I fixed the code in
MathQuizGame.somethingHappened() so that the parser thinks that the
text “c 10 0 +-x 2” is actually “/c 10 0 +-x 2” so it will return
{“10”,”0”,”+-x”,”2”}.

Also, it is easy to register a command into the command parser, so that
it is recognized as a command. I did this for my sake because I did not
want to keep writing if statements every time I added a new command.